+++ Security/ManageVoter.php
@@ -12,6 +12,7 @@ namespace c975L\UserBundle\Security;
 use Symfony\Component\Security\Core\Authentication\Token\TokenInterface;
 use Symfony\Component\Security\Core\Authorization\Voter\Voter;
 use Symfony\Component\Security\Core\Authorization\AccessDecisionManagerInterface;
+use Symfony\Component\Security\Core\User\AdvancedUserInterface;
 use Symfony\Component\DependencyInjection\ContainerInterface;
 use c975L\ConfigBundle\Service\ConfigServiceInterface;
 
@@ -106,7 +107,7 @@ class ManageVoter extends Voter
     protected function supports($attribute, $subject)
     {
         if (false !== $subject) {
-            return ($subject instanceof \Symfony\Component\Security\Core\User\AdvancedUserInterface) && in_array($attribute, self::ATTRIBUTES);
+            return ($subject instanceof AdvancedUserInterface) && in_array($attribute, self::ATTRIBUTES);
         }
 
         return in_array($attribute, self::ATTRIBUTES);